requirements.txt
# Web Framework and API
Flask==2.3.3
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
python-socketio==5.9.0
eventlet==0.33.3
fastapi==0.104.1

# Image Processing
Pillow==10.0.1

# Windows COM Interface (PowerPoint)
pywin32==306

# HTTP Client
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1

# Object Storage
minio==7.2.0

# Configuration Management
PyYAML==6.0.1

# Logging
loguru==0.7.2

# Type Hints
typing-extensions>=4.9.0

# Async Support (built-in asyncio for Python 3.7+)